refactor: move constants to seperate file
