+++ src/lib/utils.ts
@@ -5,8 +5,6 @@ import path from "node:path";
 import process from "node:process";
 import { URL } from "node:url";
 
-export const APP_NAME = "ao-deploy";
-
 /**
  * Initializes a default Arweave instance.
  */
@@ -186,14 +184,6 @@ export function parseUrl(
   return value;
 }
 
-export const defaultServices = {
-  gatewayUrl: "https://arweave.net",
-  cuUrl: "https://cu.ao-testnet.xyz",
-  muUrl: "https://mu.ao-testnet.xyz"
-};
-
-export const aoExplorerUrl = "https://www.ao.link";
-
 export function jsonStringify(value?: any): string {
   try {
     return JSON.stringify(value, null, 2);
@@ -253,21 +243,3 @@ export const getUserPkgManager: () => PackageManager = () => {
   // Default to npm if nothing else is detected
   return "npm";
 };
-
-export const AOS_QUERY = `query ($owners: [String!]!, $names: [String!]!) {
-    transactions(
-      first: 1,
-      owners: $owners,
-      tags: [
-        { name: "Data-Protocol", values: ["ao"] },
-        { name: "Type", values: ["Process"]},
-        { name: "Name", values: $names}
-      ]
-    ) {
-      edges {
-        node {
-          id
-        }
-      }
-    }
-  }`;
